requirements.txt
pymupdf
faiss-cpu
sentence-transformers
pdfplumber
groq
python-dotenv
